add attendace logic

+++ src/main/java/com/Company/SMS/DTO/Dashboard/DashboardRes.java
@@ -3,8 +3,6 @@ package com.Company.SMS.DTO.Dashboard;
 import com.Company.SMS.DTO.Attendance.AttendanceChartRES;
 import com.Company.SMS.DTO.Exam.ExamRES;
 import com.Company.SMS.DTO.Grades.GradeAverageRES;
-import com.Company.SMS.DTO.Student.StudentRES;
-import com.Company.SMS.DTO.Student.TopStudentRES;
 import com.Company.SMS.entities.Attendance;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -16,7 +14,7 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class DashboardRes {
-    private double absentForStudents;
+    private Long absentForStudents;
     private Long countExams;
     private Long totalTopStudents;
     private Long totalLowStudents;
@@ -24,3 +22,4 @@ public class DashboardRes {
     private List<AttendanceChartRES> attendanceChart;
     private List<ExamRES> exams;
 }
+
